requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.13.1
    # via -r requirements.in
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via starlette
click==8.1.7
    # via uvicorn
exceptiongroup==1.2.0
    # via anyio
fastapi==0.109.2
    # via -r requirements.in
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via uvicorn
idna==3.6
    # via anyio
mako==1.3.2
    # via alembic
markupsafe==2.1.5
    # via mako
psycopg2-binary==2.9.9
    # via -r requirements.in
pydantic==2.6.1
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.16.2
    # via pydantic
pydantic-settings==2.2.1
    # via -r requirements.in
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   pydantic-settings
sniffio==1.3.0
    # via anyio
sqlalchemy[asyncio]==2.0.27
    # via
    #   -r requirements.in
    #   alembic
starlette==0.36.3
    # via fastapi
typing-extensions==4.9.0
    # via
    #   alembic
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   uvicorn
uvicorn==0.27.1
    # via -r requirements.in
